reorganized some files into the right folders
